% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cces-sample.R
\docType{data}
\name{acs_GA}
\alias{acs_GA}
\title{ACS Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 980 rows and 8 columns.
}
\usage{
acs_GA
}
\description{
Post-stratification dataset for \code{cces_GA}.
}
\keyword{datasets}
