% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cces-sample.R
\docType{data}
\name{elec_GA}
\alias{elec_GA}
\title{Election outcome (often the quantity to estimate)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 3 columns.
}
\usage{
elec_GA
}
\description{
Election outcome (often the quantity to estimate)
}
\keyword{datasets}
