% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{error_lbl}
\alias{error_lbl}
\title{Labels for accuracy metric}
\usage{
error_lbl(
  truth,
  estimate,
  show_metrics = c("rmse", "mean"),
  metrics_lbl = c(rmse = "RMSE", mean = "MAD"),
  pp_accuracy = 0.1
)
}
\arguments{
\item{truth}{Vector of true values}

\item{estimate}{Vector of estimates, must be the same length as \code{truth}.
In fact, the metrics are invariant to which goes in which.}

\item{show_metrics}{The metrics to show. Defaults to RMSE and accuracy}
}
\description{
Labels for accuracy metric
}
