% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brms.R
\name{fit_brms_binomial}
\alias{fit_brms_binomial}
\title{Fit a brms model with count data}
\usage{
fit_brms_binomial(
  .formula,
  .data,
  verbose = TRUE,
  .prior = c(prior_string("normal(0, 1)", class = "b"), prior_string("normal(0, 1)",
    class = "sd"), prior_string("normal(0, 1)", class = "Intercept")),
  .iter = 2000,
  .warmup = floor(.iter/2),
  .cores = 4,
  .chains = 4,
  .threads = NULL,
  .backend = "cmdstanr",
  .seed = 2138
)
}
\arguments{
\item{.formula}{model specification}

\item{.data}{collapsed survey dataset, built from ccesMRPprep::build_counts}

\item{verbose}{Whether to show iteration messages}

\item{.prior}{prior specification that can be interpreted by brms. The default
is a standard normal prior, which is tighter than the brms default but has
shown to have good prior posterior draws}

\item{.iter}{Number of total iterations.}

\item{.warmup}{Of the iterations, how much are burn-ins. Defaults to half.}

\item{.cores}{Number of cores to uses}

\item{.chains}{Number of chains to pass on fit_brms}

\item{.threads}{The number of threads to do within-chain parallelization. Defaults
to not using, which is NULL}

\item{.backend}{The backend argument of brms. Currently only \code{"cmdstanr"}}

\item{.seed}{seed for randomization to pass into brm}
}
\description{
Fit a brms model with count data
}
