% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cces-sample.R
\docType{data}
\name{fit_GA}
\alias{fit_GA}
\title{Fitted brms model}
\format{
An object of class \code{brmsfit} of length 20.
}
\usage{
fit_GA
}
\description{
A fitted multilevel of the formula
\code{response ~ (1|age) + (1 + female |educ) + clinton_vote + (1|cd)}
where \code{response} is a binary variable for voting for Hilary Clinton in
2016. The sample is 2062 respondents in Georgia in 2016.
}
\keyword{datasets}
