% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-adjust.R
\name{invlogit}
\alias{invlogit}
\title{Traditional inverse logit. Turn logit scale to probability.}
\usage{
invlogit(x)
}
\arguments{
\item{x}{input}
}
\description{
Traditional inverse logit. Turn logit scale to probability.
}
\keyword{internal}
