% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-step.R
\name{poststrat_onestep}
\alias{poststrat_onestep}
\title{Do post-stratification and summarize draws in one step}
\usage{
poststrat_onestep(
  .formula,
  .data,
  poststrat_tgt,
  area_var = "cd",
  count_var = "count",
  ...
)
}
\arguments{
\item{.formula}{Formula in \verb{binary y ~ (1|x1) + (1|x2)} form.}

\item{.data}{Individual-level dataset}

\item{poststrat_tgt}{The poststratification target. It must contain the column
\code{count}, which is treated as the number of \code{trials} in the binomial model.}

\item{area_var}{Character for the variable(s) that corresponds to the area to
aggregate to.}

\item{count_var}{A character string for the variable name for the population
count in the \code{poststrat_tgt} dataframe. This will be renamed as if it is
a trial count in the model. Defaults to \code{"count"}.}

\item{...}{Additional arguments to pass to the model fitting function, \code{fit_brms()}}
}
\description{
Do post-stratification and summarize draws in one step
}
\details{
A simple wrapper around \code{poststrat_draws} and \code{summ_sims}.
}
