% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-adjust.R
\name{posthoc_error}
\alias{posthoc_error}
\title{Compute absolute deviation}
\source{
https://github.com/Catalist-LLC/unemployment/blob/master/unemployment_cps_mrp/helper_functions/GetYHat.R
}
\usage{
posthoc_error(delta, y, n, yhat)
}
\arguments{
\item{delta}{The parameter of interest}

\item{y}{The true target}

\item{yhat}{A vector of current estimate}
}
\description{
Compute absolute deviation
}
\author{
Yair Ghitza
}
