% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-adjust.R
\name{posthoc_intercept}
\alias{posthoc_intercept}
\title{Find intercept correction for cell estimates}
\source{
https://github.com/Catalist-LLC/unemployment/blob/master/unemployment_cps_mrp/helper_functions/GetYHat.R
}
\usage{
posthoc_intercept(y, n, yhat)
}
\description{
For a given geography g, there may be a value y_g which is
the ground truth. You have C cells with estiamtes of y that
may be biased. This function, prosed by Ghitza and Gelman
and Ghitza, will find a intercept shift for all C cells
to best fit the estimand y. It is the argmin of the sum of
absolute values of the deviation.
}
\examples{
 cell_sims <- poststrat_draws(fit_GA, poststrat_tgt = acs_GA)
 mrp_by_edu <- summ_sims(cell_sims, area_var = c("cd", "educ"))

}
\author{
Yair Ghitza
}
