% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_poststrat.R
\name{poststrat_draws}
\alias{poststrat_draws}
\title{Get MCMC draws of post-stratified estimate of demog x cd cells}
\usage{
poststrat_draws(model, poststrat_tgt, orig_data)
}
\arguments{
\item{model}{stan model from fit_brms}

\item{poststrat_tgt}{The poststratification target. It must contain the column
\code{count}, which is treated as the number of \code{trials} in the binomial model.}

\item{orig_data}{original survey data, only used to subset the poststratification,
containing question data.}
}
\value{
A tidy dataset with \code{qID} x \code{cd} x \code{iter} number of rows,
where \code{qID} is the number of questions (outcomes), \code{cd} is
the number of geographies, and \code{iter} is the number of iterations estimated in
the MCMC model.  The demographic cells within a district are averaged across,
and a MRP estimate is computed.
It contains the columns
\describe{
\item{iter}{The number of iterations}
\item{cd}{The geography}
\item{qID}{The question}
\item{p_mrp_est}{The proportion of success, estimated by MRP.}
}
}
\description{
Get MCMC draws of post-stratified estimate of demog x cd cells
}
