% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cces-sample.R
\docType{data}
\name{summ_GA}
\alias{summ_GA}
\title{Summary statistics of posterior samples.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 7 columns.
}
\usage{
summ_GA
}
\description{
Summary statistics of posterior samples.
}
\examples{
# Shorthand for
drw_GA <- poststrat_draws(fit_GA, poststrat_tgt = acs_GA)
summ_GA <- summ_sims(drw_GA)
}
\keyword{datasets}
