% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brms.R
\name{fit_brms}
\alias{fit_brms}
\title{Fit a binomial brms model}
\usage{
fit_brms(
  .formula,
  .data,
  .prior = c(prior_string("normal(0, 1)", class = "b"), prior_string("normal(0, 1)",
    class = "sd"), prior_string("normal(0, 1)", class = "Intercept")),
  prior_PD = FALSE,
  .seed = 2138
)
}
\arguments{
\item{.formula}{model specification}

\item{.data}{collapsed survey dataset, built from ccesMRPprep::build_counts}

\item{.prior}{prior specification that can be interepreted by brms. The default
is a standard normal prior, which is tighter than the brms default but has
shown to have good prior posterior draws}

\item{prior_PD}{whether to etimate the prior predictive distribution}

\item{.seed}{seed for randomization 02138
Fit brms regressions based on input data}
}
\description{
Fit a binomial brms model
}
