% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brms.R
\name{fit_brms_binomial}
\alias{fit_brms_binomial}
\title{Fit a brms model with count data}
\usage{
fit_brms_binomial(
  .formula,
  .data,
  .prior = c(prior_string("normal(0, 1)", class = "b"), prior_string("normal(0, 1)",
    class = "sd"), prior_string("normal(0, 1)", class = "Intercept")),
  .iter = 2000,
  .warmup = floor(.iter/2),
  .cores = 4,
  .chains = 4,
  verbose = TRUE,
  .seed = 2138
)
}
\arguments{
\item{.formula}{model specification}

\item{.data}{collapsed survey dataset, built from ccesMRPprep::build_counts}

\item{.prior}{prior specification that can be interpreted by brms. The default
is a standard normal prior, which is tighter than the brms default but has
shown to have good prior posterior draws}

\item{.iter}{Number of total iterations.}

\item{.warmup}{Of the iterations, how much are burn-ins. Defaults to half.}

\item{verbose}{Whether to show iteration messages}

\item{.seed}{seed for randomization to pass into brm
Fit brms regressions based on input data}
}
\description{
Fit a brms model with count data
}
