% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_util.R
\name{logit_ghitza}
\alias{logit_ghitza}
\title{A logit transformation as implemented by Ghitza}
\usage{
logit_ghitza(x, digits = 5)
}
\arguments{
\item{x}{input}

\item{digits}{The number of digits to round by. Larger values indicate
more closer to a theoretical inverse logit.}
}
\description{
This logit transformation (from a probability scale to unbounded) does some
trimming around extreme values.
}
\examples{
 logit <- function(x) log(x/(1 - x))
 logit(0.000001)
 logit_ghitza(0.000001, digits = 5) # the default
 logit_ghitza(0.000001, digits = 1)
 logit_ghitza(0.000001, digits = 10)

}
\seealso{
posthoc_error posthoc_intercept
}
\author{
Yair Ghitza
}
