% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-adjust.R
\name{logit_ghitza}
\alias{logit_ghitza}
\title{A logit transformation as implemented by Ghitza}
\usage{
logit_ghitza(x, digits = 5)
}
\description{
This logit transformation (from a probability scale to unbounded) does some
trimming around extreme values.
}
\examples{
 arm::logit(0.000001)
 logit_ghitza(0.000001, digits = 5) # the default
 logit_ghitza(0.000001, digits = 1)
 logit_ghitza(0.000001, digits = 10)

}
\seealso{
posthoc_error posthoc_intercept
}
\author{
Yair Ghitza
}
