% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-step.R
\name{mrp_onestep}
\alias{mrp_onestep}
\title{Fits and tidies MRP outputs in one step}
\usage{
mrp_onestep(
  .formula,
  .data,
  poststrat_tgt,
  area_var = "cd",
  weight_var,
  add_on = NULL,
  ...
)
}
\arguments{
\item{add_on}{Any area-level data to be merged with the output,
for example validation data}
}
\description{
Fits and tidies MRP outputs in one step
}
\details{
Combines \code{fit_brms}, \code{poststrat_draws}, and \code{direct_est}. See \code{scatter_45}
for options on visualization
}
\examples{
\dontrun{
mrp_fit <- mrp_onestep(response ~ (1|educ) + (1|cd),
                       .data = cces_GA,
                       poststrat_tgt = acs_GA,
                       area_var = "cd",
                       weight_var = "weight_post",
                       add_on = distinct(acs_GA, cd, clinton_vote))

scatter_45(mrp_fit, clinton_vote, p_mrp_est,
           xlab = "Clinton Vote", ylab = "MRP Estimate")
}

}
