% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_1way.R
\name{posthoc_intercept}
\alias{posthoc_intercept}
\title{Find intercept correction for cell estimates}
\source{
FindDelta function at https://github.com/Catalist-LLC/unemployment/blob/master/unemployment_cps_mrp/helper_functions/GetYHat.R
}
\usage{
posthoc_intercept(xi, ests, n, search = c(-5, 5))
}
\arguments{
\item{xi}{The true target}

\item{ests}{A vector of current estimate, by cell}

\item{n}{The sample size of the estimates, by cell}

\item{search}{The lower and upper endpoints of the interval to search}
}
\value{
The value of delta or the intercept that minimizes
the absolute deviation in total. The value is on the logit scale.
To translate to a probability, use invlogit.
}
\description{
For a given geography g, there may be a value y_g which is
the ground truth. You have C cells with estimates of y that
may be biased. This function, proposed by Ghitza and Gelman
and Ghitza, will find a intercept shift for all C cells
to best fit the estimand y. It is the argmin of the sum of
absolute values of the deviation.
}
\examples{

 biased_ests <- ccesMRPrun:::invlogit(rnorm(n = 100, mean = 1, sd = 1))
 sizes <- rbinom(n = 100, size = 100, prob = 0.1)
 tru <- 0.5

 posthoc_intercept(tru, biased_ests, sizes)

}
\seealso{
posthoc_error
}
\author{
Yair Ghitza
}
