% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_2way.R
\name{posthoc_twoway}
\alias{posthoc_twoway}
\title{Two-way Calibration}
\source{
Kuriwaki, S., Ansolabehere, S., Dagonel, A., & Yamauchi, S. (2021).
The Geography of Racially Polarized Voting: Calibrating Surveys at the
District Level. \url{https://doi.org/10.31219/osf.io/mk9e6}
}
\usage{
posthoc_twoway(
  data,
  var_area,
  var_group,
  tgt_area,
  tgt_group,
  X,
  n_area,
  n_group,
  n_total,
  delta_init = NULL
)
}
\arguments{
\item{data}{Estimates stored in the long format. Must be coercible to a non-tibble dataframe.
The column should be named \code{est} and the sample size should be called \code{n_gj}.}

\item{var_area}{Variable name (char) for area in \code{data}.}

\item{var_group}{Variable name (char) for group in \code{data}.}

\item{tgt_area}{Vector of true values for area.}

\item{tgt_group}{Vector of true values for group.}

\item{X}{Design matrix. E.g., \code{data.matrix(~cd+race-1, data = data)}.}

\item{n_area}{Vector consists of population sizes in each area.}

\item{n_group}{Vector consists of population sizes for each group.}

\item{n_total}{Scalar of total number of population.}

\item{delta_init}{Initial values of delta.}
}
\value{
Data frame with new columns \code{"est_corrected"} and \code{"delta"}
}
\description{
Two-way Calibration
}
\examples{
# Single estimate


}
