% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brms.R
\name{prior_pd_binomial}
\alias{prior_pd_binomial}
\title{Use priors and covariates to generate a prior predictive distribution}
\usage{
prior_pd_binomial(
  .formula,
  .data,
  .prior = c(prior_string("normal(0, 1)", class = "b"), prior_string("normal(0, 1)",
    class = "sd"), prior_string("normal(0, 1)", class = "Intercept")),
  .seed = 2138
)
}
\arguments{
\item{.formula}{model specification}

\item{.data}{collapsed survey dataset, built from ccesMRPprep::build_counts}

\item{.prior}{prior specification that can be interepreted by brms. The default
is a standard normal prior, which is tighter than the brms default but has
shown to have good prior posterior draws}

\item{.seed}{seed for randomization to pass into brm
Fit brms regressions based on input data}
}
\description{
Use priors and covariates to generate a prior predictive distribution
}
