% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{scatter_45}
\alias{scatter_45}
\title{Standard scatterplot with 45 degree line}
\usage{
scatter_45(
  tbl,
  xvar,
  yvar,
  lblvar = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  size.point = 0.8,
  ubvar = NULL,
  lbvar = NULL,
  by_form = NULL,
  by_labels = NULL,
  show_error = TRUE,
  expand_axes = TRUE,
  ...
)
}
\arguments{
\item{tbl}{dataset of summary statistics}

\item{xvar}{Variable to put on x-axis, unquoted}

\item{yvar}{Variable to put on y-axis, unquoted}

\item{lblvar}{Variable to use as labels for \code{geom_text_repel}, unquoted}

\item{xlab, ylab}{x and y-axis labels, respectively}

\item{xlim, ylim}{x and y-axis limits, respectively}

\item{size.point}{Size of points to use in ggplot}

\item{ubvar, lbvar}{Variable to use as upper and lower bounds for \code{geom_errorbar}, unqoted}

\item{by_form}{If the dataset is in long form with separate rows for different
model estimates, you can supply a formula to be passed on to \code{facet_wrap()} to
have separate facets for each model.}

\item{by_labels}{A named vector for the facets, where the names are the names
of the unique values of the variable by specified in \code{by_form} (e.g. "model") and
the values are the corresponding characters to recode to.}

\item{show_error}{Whether or not to show the accuracy metrics in caption}

\item{expand_axes}{Whether to expand the axes so that the plot is a square,
even if there is more whitespace. Overrides xlim and ylim.}

\item{...}{Additional arguments sent to the \code{error_lbl} function}
}
\description{
Standard scatterplot with 45 degree line
}
\examples{
library(dplyr)

mrp_df <- summ_sims(poststrat_draws(fit_GA, poststrat_tgt = acs_GA)) \%>\%
  left_join(elec_GA)


scatter_45(mrp_df,
           clinton_vote,
           p_mrp_est,
           lblvar = cd,
           lbvar = p_mrp_050,
           ubvar = p_mrp_950,
           xlab = "Clinton Vote",
           ylab = "MRP Estimate")


}
