% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{scatter_45}
\alias{scatter_45}
\title{Standard scatterplot with 45 degree line}
\usage{
scatter_45(
  tbl,
  xvar,
  yvar,
  lblvar = NULL,
  xlab = NULL,
  ylab = NULL,
  show_error = TRUE,
  ...
)
}
\arguments{
\item{tbl}{dataset of summary statistics}

\item{xvar}{Variable to put on x-axis, unquoted}

\item{yvar}{Variable to put on y-axis, unquoted}

\item{xlab, ylab}{X and y-axis labels, respectivey}

\item{show_error}{Whether or not to show the accuracy metrics in caption}

\item{...}{Additional arguments sent to the \code{error_lbl} function}
}
\description{
Standard scatterplot with 45 degree line
}
