% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_draws.R
\name{summ_sims}
\alias{summ_sims}
\title{Get summary statistics from draws (of counts)}
\usage{
summ_sims(sims)
}
\arguments{
\item{sims}{output of \link{poststrat_draws}}
}
\description{
Get summary statistics from draws (of counts)
}
